Created a useCurrentMeasure hook for use inside the edit measure components Eliminated the EditMeasureRoutes component
