+++ src/components/EditMeasure/MeasureDetails/MeasureDetails.tsx
@@ -1,11 +1,10 @@
 import React, { useEffect, useState } from "react";
 import tw from "twin.macro";
 import { Route, Switch, useRouteMatch } from "react-router-dom";
-import Measure from "../../../models/Measure";
-import useMeasureServiceApi from "../../../api/useMeasureServiceApi";
 import MeasureInformation from "./MeasureInformation";
 import MeasureSteward from "./MeasureSteward";
 import MeasureDetailsSidebar from "./MeasureDetailsSidebar";
+import useCurrentMeasure from "../useCurrentMeasure";
 
 interface MeasureParam {
   id: string;
@@ -15,21 +14,9 @@ const Grid = tw.div`grid grid-cols-4 gap-4 ml-6`;
 const Content = tw.div`col-span-3`;
 
 export default function EditMeasure(params: MeasureParam) {
-  const { id } = params;
-  const [measure, setMeasure] = useState<Measure>(null);
-  const [loading, setLoading] = useState<boolean>(true);
-  const measureServiceApi = useMeasureServiceApi();
   const { url } = useRouteMatch();
   const stewardLink = `${url}/measure-steward`;
-
-  useEffect(() => {
-    measureServiceApi.fetchMeasure(id).then((value: Measure) => {
-      setMeasure(value);
-      setLoading(false);
-    });
-  }, [measureServiceApi, id]);
-
-  const loadingDiv = <div>Loading...</div>;
+  const measure = useCurrentMeasure();
 
   const links = [
     {
@@ -42,7 +29,7 @@ export default function EditMeasure(params: MeasureParam) {
     },
   ];
 
-  const contentDiv = (
+  return (
     <>
       <Grid>
         <MeasureDetailsSidebar links={links} />
@@ -59,6 +46,4 @@ export default function EditMeasure(params: MeasureParam) {
       </Grid>
     </>
   );
-
-  return loading ? loadingDiv : contentDiv;
 }
